requirements.txt
notebook==7.0.4
openai==0.28.1
Pillow==9.5.0
wheezy.captcha==3.2.0
pytesseract==0.3.10
Levenshtein==0.22.0
mss==9.0.1
Kivy==2.2.1
mouse==0.7.1
opencv-python==4.8.1.78

